LogDir: cleanup
getPreviousLogdir[s]() had the side effect of modifying the log dir
used by the instance. In many cases, the parameter was the one
from the SyncContext instance that the LogDir had access to.

This patch makes this a bit more obvious by using the SyncContext
log dir the default and only calling LogDir::setLogdir() explicitly
when needed.