command line: compare against dump in last session involving source (MB #7708)
This patch adds more intelligent searching for the ".after" database
dump of each source that gets compared. This solves the problem that
if the last session involved only one source, comparison of all other
sessions was called "impossible".

The patch is careful to pick the last session which really involved
the source. This is not based on the existence of the dump, because
writing that might have failed or been disabled. Instead the status.ini
of the session is looked at. Note that this is different from the search
in dumpDatabases(), which can work with any dump of the source in
any peer at any time.

As a minor clean-up, some variables were named to reflect what
they really refer to nowadays (dumps are done in directories, not files).