SHA-256: use glib or Mozilla NSS
Stronger hash algorithms are necessary for the identification of
file content via a hash value (same approach as in git). SHA-256
is taken from glib if glib is already a SyncEvolution dependency
for other reasons, otherwise the configure script looks for
Mozilla NSS.

Another candidate is libgcrypt, but there's no code for that yet.

All of these libraries are LGPL or more liberal, so they don't
impose any additional requirements on users of SyncEvolution,
in contrast to OpenSSL (advertisement clause!).