autotools: some more fixes
Changes done a while ago, checking in now...

"make install" with the plugin off still tried to create the plugin directory.
I went for excluding entering the sub directory, as in the original patch, in
particular because DIST_SUBDIRS had to be updated anyway. Apparently when it is
set, it is used by "make dist" instead of SUBDIRS. I thought "make dist" would
use a union of the two.

update-samples.pl was missing, broke "make distcheck".