ObexTransportAgent: error handling
When the transport returns LINK_ERR with a response code = 0, this
is not an error for some phones (as tested by 7210c), however we also
found for N900 it is a real transport error.
Since error case is rare and we can finally catch it (just some confusing error
states), I will prefer to treat this case as normal.