virtual source in client: tell engine about superdatastore and URI
Super datastores do not have a target entry in a binfile client's profile.
They have to be parameterized via their sub-datastores, using the
special <super>uri format for the sub-datastores' URI. "super" is
the name of the super datastore and "uri" its URI. The sync mode
is also taken from the sub datastores.