SyncSource::Operations: added callback for starting to use source
The new m_startAccess is called before the source's m_startDataRead
and thus marks the point where a source is used for the first time
during a sync. This is necessary for initializing sources in a server
on demand.

Because this happens inside SyncContext, at least callbacks must
be possible for outside classes. Setting the other operations is
still limited to derived classes.