SyncML server: delayed checking of sources (MB #7710)
With this patch, SyncML server sources are only opened() and their
data dumped when a client really uses them. As before, sources are
only enabled in the server if their sync mode is not "disabled". This
tolerates sources which cannot be instantiated because their "type" is
not supported.

The patch changes the SourceList and its methods so that they can do
the database dumps and comparisons for a single source at a
time. SourceList tracks which of its sources were dumped before the
sync and uses that information at the end to produce the "after sync"
comparison.

That "after sync" comparison was a reduced copy of the
dumpLocalChanges() source code. The copy was replaced with a suitably
parameterized call to dumpLocalChanges(), which became easy after
adding the "oldSession" parameter in a recent patch. That output now
is as follows:

-------------------------> snip <-----------------------------------

Changes applied during synchronization:
+---------------|-----------------------|-----------------------|-CON-+
|               |         LOCAL         |        REMOTE         | FLI |
|        Source | NEW | MOD | DEL | ERR | NEW | MOD | DEL | ERR | CTS |
+---------------+-----+-----+-----+-----+-----+-----+-----+-----+-----+
|   addressbook |  0  |  0  |  0  |  0  |  0  |  0  |  0  |  0  |  0  |
|   two-way, 0 KB sent by client, 0 KB received                       |
|   item(s) in database backup: 20 before sync, 20 after it           |
+---------------+-----+-----+-----+-----+-----+-----+-----+-----+-----+
|      calendar |  0  |  0  |  0  |  0  |  0  |  0  |  0  |  0  |  0  |
|      two-way, 0 KB sent by client, 0 KB received                    |
|      item(s) in database backup: 20 before sync, 20 after it        |
+---------------+-----+-----+-----+-----+-----+-----+-----+-----+-----+
|          start Wed Feb 10 16:38:15 2010, duration 0:02min           |
|               synchronization completed successfully                |
+---------------+-----+-----+-----+-----+-----+-----+-----+-----+-----+

Data modified locally during sync:
*** addressbook ***
no changes
*** calendar ***
no changes

-------------------------> snip <-----------------------------------

Previously the last heading was "Changes applied to client during
synchronization", which is wrong for the server (it is not a
client) and did not properly distinguish between item and data
changes (items may be changed without affecting the set of data,
as in removing one item and adding it with the same content).

In a server, the "*** <source> ***" part is only printed for active
sources, whereas the table always contains all sources with sync mode
!= "disabled". If we had progress events for the server, it should be
more obvious that some sources were not really used during the
sync. Alternatively we could also remove them from the report.

Also fixed several other such "to server/client" messages. They were
written from the perspective of a client and were wrong when running
as server. Using "remotely" and "locally" instead works on both client
and server.