slow sync detection + virtual source: fixed user message
When an unwanted slow sync is detected for one or more normal sources
which are part of a virtual sync source, then the user must rerun
the sync with the virtual source named on the command line.

To implement this, a set of source names is more suitable than a list.
Changed SyncML::slowSyncExplanation() accordingly. The information
that a source is part of a virtual source is recorded in the new
SyncSourceReport::m_virtualSource. It is checked first before adding
the normal source's name to the set.

Saving the virtual source name in the status.ini and sending it via
D-Bus GetReport() is perhaps not necessary, but just in case it was
also added under the key "source-<source>-virtualsource = <virtual source>".