libbluetooth compatibility layer (MB #9289)
libbluetooth.so.2 and libbluetooth.so.3 can be accessed via
dynamically determined function pointers. This removes the hard
dependency on libbluetooth. ObexTransportAgent will refuse to do the
channel search if no suitable version is found; manually specifying
the channel is a workaround.

More importantly, the functions which changed between 2 and 3 can be
called via the extended prototypes, with wrappers falling back to the
less secure version 2 implementations if that is all that is
available.