virtual datastore: allow alias (MB #9664)
When running with "calendar", "todo" both inactive and only activated
indirectly via "calendar+todo", the XML config listed <datatores>
called "calendar" and "todo" (with aliases, if specified) and the
"calendar+todo" <superdatastore>.

The N85 can start a sync with a config like that, but it sends events
and tasks to the "./Calendar" URI, which then end up the calendar
<datastore> instead of getting dispatched via the <superdatastore>.

First problem: setting an alias for <superdatastore> was
missing. Fixed in this patch.

Second problem: the engine might still pick the normal <datastore>
when asked for "./Calendar". Not fixed in this patch. Need to rename
sub-datastores?