virtual sources: avoid name collisions with normal sources (MB #9664)
As suspected in the previous commit for this issue, the engine still
picks the normal <datastore> when asked for "./Calendar". This patch
renames, for example, "calendar" to "calendar+todo@calendar" to avoid
cluttering the global namespaces for URIs with sources that clients
should not synchronize against.

The "@" separator is used because it is already a reserved character
for config names, albeit not yet for source names. A slash "/" would
have been nicer because it cannot be part of a source name, but when
trying that the Synthesis engine treated the second part as URI
parameters (instead of part of the source name).