SyncSource API: added m_isEmpty operation (MB #7708)
The current solution of checking for empty sources by looking at the
"before" database dump is a hack which fails if creating that backup
was disabled. It also prevents delaying the dump until after the
start of the session, because the information must be available
before starting the session, so that the XML config can be created
with slow sync check enabled if needed.

This patch adds a new operation which backends must implement if they
want to provide the "is empty" information for the slow sync
detection. This is optional. If the information is not available, slow
sync detection is always on, even when not strictly needed.