TrackingSyncSource: added isEmpty() pure virtual method (MB #7708)
This implements the new m_isEmpty operation with a pure virtual
TrackingSyncSource::isEmpty() callback. It follows the philosophy
that TrackingSyncSource should
a) hide the underlying operation mechanism and
b) prevent derived classes from compiling unless they do
   everything that is expected of them.

All of the in-tree backends were adapted so that they provide at least
a dummy implementation of the call, to keep them in a usable state.

The file backend implementation is okay. The Evolution backends get the
job done correctly, but not efficiently (MB #9750). Maemo and XMLRPC
backends always return false, which is acceptable but should be changed
(disables the "allow slow sync for empty databases" heuristic).