SyncContext.cpp: added testing of session handling
These tests cover the LogDir::haveDifferentContent()
utility function and the backup feature of sync sessions.
The latter is based on instantiating a SourceList with
FileSyncSources which can be pointed to different sets of
items via a symlink. This allows to simulate changes between
sessions as well as during a session.