LogDir::startSession(): fixed collision check
The code was meant to ensure that whenever multiple session,
regardless of which peer, happen to fall on the same session
(unlikely, but happens in testing), an increasing sequence
number is attached to the new directory name.

This only worked if the current peer was among those who started in
that second, otherwise the collision check was avoided. In addition,
the code recycled sequence numbers of directories which had been
removed. So in the even more unlikely case of doing the following all
in one second, we ended up reusing the base name and thus had
an incorrect ordering:
- create foo-2010-01-31-12-00
- create foo-2010-01-31-12-00-1
- delete 2010-01-31-12-00
- reuse foo-2010-01-31-12-00

This found when adding unit testing of the revised expiry mechanism.

The underlying problem is that sorting of sessions depends on linear
time. Because we use local time for the directory names and clocks might
get reset, this is not guaranteed (MB #9759).