SyncContext: fixed broken virtual source data format check
First, the check was only done when the user forced a type. I'm
not sure why it was limited like that. IMHO alerting the user of
an inconsistent config would make sense also for other cases.

Second, it didn't work correctly because it compared against
getPeerMimeType(). This is not the type configured in the config, but
rather something hard-coded in the backend. For
"type=todo:text/x-vcalendar" it returned "text/calendar", which led to
a warning although the virtual "calendar+todo" also used
"text/x-vcalendar".

Third, the warning was only printed once and spelled inconsistently.
If both sub data sources are configured incorrectly, SyncEvolution
should complain about both before proceeding. It might also abort,
but that is debatable and not changed in this patch.