maxMsgSize: increased from 20000 to 150000
The previous message size was to small to hold large DevInf structures.
In contrast to items, the DevInf cannot be split between messages.
There was no particular reason for the small size, so increasing the
size shouldn't introduce new problems. It is still small enough to
be buffered in memory.