deviceId: use syncevolution- prefix, helps Horde (MB #9347)
Previously we used "sc-pim-" as prefix for our device ID because
my.FUNAMBOL depended on it for distinguishing phones from other
devices. It supported only one phone, but multiple others. Without
the right prefix it would do slow syncs when switching between
different clients.

This behavior was changed, which allows us to use our own prefix.
That solves a problem with Horde, which treated our client like
an old, broken Funambol client because of the prefix and then
sent us calendar items in the old Funambol SIF-C format instead
of vCalendar/iCalendar.