Evolution Address Book: avoid picking CouchDB by default, again (MB #7877)
The first attempt at solving this problem was incomplete. In addition
to reordering the databases so that CouchDB databases are listed last
(previous patch), we also must change the code so that the first
database *after* reordering is marked as default *and* that actual
syncs use that database.

Therefore this patch changes EvolutionContactSource::getDatabases() so
that the setting of m_isDefault is done at the end and
EvolutionSyncSource::findSource() so that it looks at that result when
no DB is selected explicitly.