syncevolution.org packages: conflict with system libs (MB #9811)
Our .deb and .rpm packages contain /usr/lib/ libraries which
are the same as packaged by distros in separate packages, for
example libsynthesis0 and libsmltk0 (Debian). This patch
makes a guess at what these packages might be based on the .so.x
file name and adds them to the "Conflicts" package attribute.