syncevo-dbus-server: accept SAN 1.0/1.1 pacakge (MB#9312)
Also accept SAN 1.1 package, this requres enabling SAN_1_1 build option for
synthesis SDK.
In dbus server, when receving a message with content type 'xml/wbxml', we can
not think it as a client contact us, it may also be possibly a server sending
SAN 1.1 package to us; in this case first take it as a SAN package and try to
parse it, if fails fall back to a normal synchronization message from a client.
The SAN parsing did not fail as expected in the original implementation, thus
a patch for synthesis sdk is essential.