EvolutionCalendarSource: change tracking when deleting a child event
When removing a child event (RECURRENCE-ID set), the behavior of EDS
for the parent event (without the RECURRENCE-ID) is
undefined. Client::Source::ical20::testLinkedItemsRemoveNormal assumed
that the parent's LAST-MODIFIED was not affected.

Nightly testing recently showed that at least some EDS versions modify
the parent when removing a child. More specifically, they add an
EXDATE to the parent for the recurrence that was removed.  This is
reasonable. The puzzling part is that this test failure was not
observed earlier. This might be because the test could be timing
dependent: if creating the parent and removing the child fall into the
same second, then the parent is not considered as modified.

This patch modifies the test so that it passes both when the parent is
updated and when it isn't.

In addition, this failure points towards another problem: if a server
like ScheduleWorld deletes a child event on a client, it first sends
an update for the parent and then the removal request for the
child. We should not report the parent as modified during the next
sync. Again, depending on the timing, this may or may not have
happened.

This is based on the assumption that the server knows about
child/parent events and does the right thing. If the parent event was
not modified on the server when removing the child, then with this
patch we are not going to fix that in the next sync. Such servers
should be fixed.

This patch therefore adds a check to testLinkedItemsRemoveNormal for
the change tracking and fixes EvolutionCalendarSource::removeItem() so
that it updates the change tracking when necessary. Because
TrackingSyncSource is not prepared to deal with a removal that has
side effects, its tracking config node must be made accessible for
this to work.