DBus Server: implement automatic sync (MB#6378)
Implement automatic sync feature. 3 scenarios are:
1) For a config enables autosync, an interval has passed.
2) Once users log in or resume and an interval has passed. Not
implemented yet.
3) Evolution data server notify any changes. Not implemented yet.

Auto sync manager is to manage all issues of auto sync. In its
initialization step, it reads all configs and check their properties.
If a configuration is enabled with automatic sync, possibly http
or obex-bt or both, one or more tasks for different URLs of each config
are added in the task map grouped by their intervals, holding configured
tasks. Tasks in the map are updated once any config is changed.

Tasks in the map are scheduled and checked when some conditions are met,
such as the interval has passed(the first scenario). Scheduled tasks are
put in the working queue to wait for doing sync. Before a scheduled task
is put in the working queue, it has to be checked whether there is
any existing same task in the working queue.

Once there is any task in the working queue, auto sync manager picks one
task from the working queue and create a session with a lower priority for
it than normal. However, at any time, there is at most one session.

If the session is scheduled and activated by dbus server, it gets
the privilege to do sync. After completion, auto sync manager picks another
task if has and creates a new session. Do this repeatedly unless the
working queue is empty.