LogDir: use a-z instead of number to make session dirs unique (MB #9759)
Using numbers 1..9..10.. as suffix had the problem that sorting
puts -10 in front of -2. Because we use hh-mm as base name, this
limited us to 9 sessions per minute, which is easily reached
when doing local testing.

By switching to -a to -z as suffix, we get 27 sessions per minute
before moving on to less nice characters (possible, but might
make file names unusuable eventually). That should be enough.
In local tests I got as far as -q.

With this change, the LogDir expire tests can be merged back
into one function.

An additional benefit is that the suffix is visually different
from the base name, which makes it a bit easier to identify
yyyy-mm-dd-hh-mm manually.