unexpected slow sync: fix result when using a single source
When a single source is active and we abort because of an unexpected
slow sync, then the overall sync result was set to the internal 20048
"abort datastore" status. This status must be translated back to
the root cause of the abort, the 22000 "unexpected slow sync" error.

The right place to do this is before recording the status in the
SyncReport and before returning it. Any status read after that from
individual sources will already have 20048 replaced with the
right status.