client-test: added interrupted sync tests with split items
The tests are expected to run with a maximum message size of 20000
used by both client *and* server - server must be configured
accordingly.

Then the Add and Update operations in the interrupted sync with
client B are done with an item that is larger than the message size
and thus depends on <MoreData> support. On resume, only the second
half is sent and the recipient must assemble the complete item.