DBusServer: start syncevo-dbus-server when users log in
create a desktop file to start syncevo-dbus-server when a user
logs in. The desktop file actually starts syncevo-dbus-server-startup.sh
instead of syncevo-dbus-server directly. The reason is the script is to
sleep 120 seconds and then start the dbus server not to increase starting
time of the system.

Don't use AC_CONFIG_FILES() to generate the desktop file and script.
That's because if a @variable@ references other variables in the *.in
file, substitutions won't work.