command line: check for invalid peer properties (MB #8048)
When using --print-config or --configure with a context as config name
(like @default), then per-peer properties do not have any effect. This
might be confusing for users, so the command line should better report
this as error and refuse to execute the operation.

As usual, "type" is a bit special because it is both shared and
per-peer.