SyncSourceAdmin: fixed storing of map item with no remote ID (MB #2425)
With our own server, the testInterruptRetryServerAdd case failed.  It
turned out that SyncSourceAdmin::mapid2entry() was generating " 1" for
empty remote ID and flags == 1, which then was stored as "1" and
misinterpreted as remote ID "1" with flags == 0 by
SyncSourceAdmin::entry2mapid().

Our config nodes cannot store leading or trailing spaces. We could
quote empty strings, but that would lead to backward compatibility
problems. Therefore this patch treats a single word in the value as
the flags and explicitly creates such values, instead of relying on
the key/value store to do that implicitly.