client-test: testInterruptRetry + loss of final server reply
testInterruptRetry* always failed when it simulated the loss of the
final server reply. In that case the server thinks that the session
has ended (because it cannot check successful delivery of its reply),
whereas the client notices the problem and suspends => unexpected slow
sync in next sync.

This patch counts the number of messages that go back and forth, then
stops testing before reaching the problematic message. Alternatively it
could disable slow sync checking for that last iteration, but then we end
up testing the server's slow sync implementation, which is a different
problem. Better keep the test focused on interrupt+resume.