libbluetooth3 compatibility (MB #9289)
With libbluetooth3 and --enable-evolution-compatibility,
searching for functions stopped because the _safe variants
were not found. As a result, syncevolution segfaulted
when trying to call str2ba() via a NULL pointer.

Now the code finds all available functions, then later
decides whether missing some is acceptable.

Note that for this to work, libbluetooth must not be referenced
explicitly by the binary. Otherwise we might end up with loading
libbluetooth.so.2 at startup and libbluetooth.so.3 dynamically
at runtime, with unpredictable results.