Create configuration automatically (MB#6175)
If we are contacted by a remote peer via SAN and there is not a configuration
already available, we should create a configuration automatically.
This patch moved part of SAN content matching to Connection.ready, because
creating a configuration needs Session support, that's not available in
Connection.process, defer it to Connection.ready is a safe point.