Disable PreventSlowSync if the sync session is triggered by a remote peer
'PreventSlowSync' was enabled, so that when a sync could not be conducted without
falling back to a slow sync, we will abort the sync and warn the user about this
unnormal condition.

However if the sync session is inited from a remote peer (e.g. Server Alerted
Sync), this kind of warning on our side has no way to convey to the remote peer;
thus causing remote peer confused about the sync failure. It is reasonable to
disable this feature in this kind of scenario because: 1) The remote peer might
not have 'PreventSlowSync' check capability and we must work with them 2) For
SyncEvolution server, this kind of checking should be implemented at server side.