Install: change 'autostart' data install directory
Auto start desktop files should be installed in the /etc/xdg/autostart,
which linux distros use. However, this hard-code directory prevents
distributions failed due to permission. So the change
is to use $sysconfdir to install desktop files to avoid this issue.
However, to enable desktop files, you have to use '--sysconfdir=/etc'
to make sure the desktop files are installed in the correct directory.