Remove support for accepting SAN 1.0/1.1 (MB #6175, MB #9312)
As mentioned in MB #6175, running as HTTP server no longer worked
because the incoming message was mistaken as SAN message.

This patch solves this by manually reverting the following patch:

commit 12bd955d88d90c91835c55c0d1e01fcbf60118ba
Author: Chen Congwu <congwu.chen@intel.com>
Date:   Tue Feb 23 17:01:37 2010 +0800

    syncevo-dbus-server: accept SAN 1.0/1.1 pacakge (MB#9312)

Automatic revert did not work because some of the code had been
modified in the meantime.

Reverting this might be the final solution, because we don't need to
support incoming SAN 1.0/1.1. Whoever wants to talk with us should
use SAN 1.2.