gen-autotools.sh: set versions suffix when code is not cleanly tagged
This patch is meant to catch the following problems in nightly builds:
- changes were made on the build machine without committing them
- "master" is checked out and happens to be more recent than the
  latest tag (development snapshots are not tagged, translations
  are committed to master by others)
- the Synthesis source was not tagged or is unclean

If any of the checks fails, +<date>+SE+<hash>[+unclean]+SYSYNC+<hash>[+unclean]
is appended to the version in the generated "configure.in" script, for
SyncEvolution (SE) and libsynthesis (SYSYNC).

The check of the tag is based on the assumption that dots in the
configure version were replaced with hyphens in the tag. Extra hyphens
in the tag may be used. "syncevolution" must be part of the tag, in
both SyncEvolution and libsynthesis. So for a version 1.0beta1, the
corresponding tags are syncevolution-1-0-beta-1 in SyncEvolution and
<something>+syncevolution-1-0-beta-1 in libsynthesis.