DBusServer: send log output to clients (MB#5042)
The console output of dbus server is useful. Thus, it is necessary
to send them to clients at the same time logging them in the
logging system.
DBusServer acts a Logger to pass logging messages to its parent
logger and send them as signals to clients.

Signal 'Output' is defined with 3 arguments: object path, log level,
output. If dbus server has active session, then the output is
session-specific, the 'object path' argument will be filled with
session object path. Otherwise, the object path of dbus server is
used instead of an empty string for it can't be used as object path.