client-test: check that suspend tests really suspend
The suspend tests would have passed even if the client never really
executed the suspend request. Now the tests verify whether the next
session really resumes. It does that except for some corner cases
(first message, last message), which is acceptable. When multiple
sources are involved, first and last message of each source
cannot be identified reliably, so in that case suspend is not checked.

The SyncReport must be available for this checking. The check() method
was made virtual to implement such checks in a derived class, but that
is neither convenient nor does it currently work because SyncOptions
copies the CheckSyncReport instance.