syncevo-http-server: include Python script in distribution
Because people asked for it, the Twisted-based HTTP server is
now installed when D-Bus support (which it depends on) is enabled.

It is installed as "syncevo-http-server" without the .py suffix
because that is an implementation detail that might change.

Because people need a way to figure out how to start it, usage
information is provided via optparse.